Crate hoover3_database

This module contains the database clients, model definitions, and various management functions. It also hosts a bunch of client methods for interacting with the various databases.

Re-exports

pub use charybdis;

pub use charybdis_macros;

Modules

client_query

This module contains the client methods for interacting with the various databases.

db_management

Core database management module that defines common interfaces and traits for different database backends. Provides the DatabaseSpaceManager trait and re-exports specific database implementations.

migrate

This module contains the database migration functions, including mirroring the Scylla schema into other databases.

models

This module contains the table definitions for all the database models.

Macros

impl_model_callbacks

Macro to implement Charybdis callbacks for a model struct.