hoover3_database::client_query
Module collections

Collection management module that provides functionality to create, update, and delete collections.

Functions

_get_all_collections 🔒

create_new_collection

Client API method to create a new collection, including all the databases, tables, search indexes and other setup that must happen. If a collection with said name exists, return that instead.

drop_collection

Client API method used to drop a collection, all databases and entries.

get_all_collections

Client API method, returns list of all collections in system. Cached for 1min. Cache gets dumped on CREATE, DELETE, MODIFY.

get_single_collection

Client API method, returns details for a single collection in the system.

update_collection

Client API method used to update collection title and description.