hoover3_database::client_query
Module datasources

Datasource management module that provides functionality to create, update, and delete datasources.

Functions

create_datasource

Create a new datasource in the given collection. Repeated calls using the same name will return the same datasource. A tokio spawn task is used to ensure the creation is finalized even if the user disconnects.

drop_datasource

Drop a datasource from the given collection. Dropping a non-existent datasource is a no-op.

get_all_datasources

Client API method, returns details for a single collection in the system.

get_datasource

Get details for a single datasource in the given collection.