hoover3_database::client_query
Module list_disk

List files and directories on disk.

Functions

data_root_path 🔒

get_path_metadata

Get metadata for a single file or directory on disk.

list_directory

List all files and directories in the given directory on disk.