hoover3_database::db_management::redis
Function with_redis_cache

pub async fn with_redis_cache<K, F, T>(
    redis_cache_id: &str,
    ttl_sec: u32,
    func: impl FnOnce(K) -> F + Send + 'static,
    key: &K,
) -> Result<T>
where
    K: Serialize + 'static + Send + for<'a> Deserialize<'a> + Clone,
    F: Future<Output = Result<T>> + 'static + Send,
    T: 'static + Send + Sync + Serialize + for<'a> Deserialize<'a>,

Execute a function with a Redis cache. If the cache is hit, the function is not executed. If the cache is not hit, the function is executed and the result is cached. It also uses a lock, so only one instance for a given key is executed at one time.