hoover3_database::db_management::redis
Function with_redis_lock

pub async fn with_redis_lock<F>(
    redis_lock_id: &str,
    func: F,
) -> Result<F::Output>
where
    F: Future + Send + 'static,
    F::Output: 'static + Send + Sync,

Acquire a distributed lock on a Redis server and execute a function while holding the lock. The lock is automatically released after the function completes. Function execution timeout is 5 days. Lock acquisition timeout is 2 minutes.