hoover3_database
Module migrate

This module contains the database migration functions, including mirroring the Scylla schema into other databases.

Modules

schema_meilisearch 🔒

schema_nebula 🔒

schema_scylla 🔒

Functions

_drop_collection 🔒

_migrate_collection 🔒

_migrate_common 🔒

_migrate_nebula_collection

API Client method to migrate the Nebula database for a collection.

drop_collection

Drop databases for a single collection.

get_collection_schema

API Client method to get all the database schemas for a collection.

get_package_dir

Sometimes we run from workspace root. Sometimes we run from package root. We need to point to correct migration dirs; so we need to identify this package’s dir. This function tries a few variations and finds the hoover3_database package root.

migrate_all

Migrate all databases for all collections.

migrate_collection

Migrate databases for a single collection.

migrate_common

Migrate the common database schema (the one not related to any collection).

nebula_get_schema

API Client method to get the Nebula database schema for a collection.

scylla_migrate_collection 🔒