hoover3_database::models
Module collection

This module contains the table definitions for all the collection related models.

Modules

_model_extra 🔒

This module implements the impl_model_callbacks macro, which is used to add Charybdis callbacks to model structs. These callbacks are used to insert/update/delete rows in the secondary databases, Nebula and Meilisearch.

_nebula_edges

This module contains the definitions for all the nebula edges. Unit structs are used to identify edges, and are converted to GraphEdgeTypes for use at runtime.

datasource

This module contains the table definitions for all the datasource related models.

filesystem

This module contains the table definitions for all the filesystem related models.

Macros

impl_model_callbacks

Macro to implement Charybdis callbacks for a model struct.

Structs

DatabaseExtraCallbacks

Extra runtime information for running database insert/update/delete operations across multiple databases. This struct consists of various database handles and schema information, to be used in row callbacks.

InsertEdgeBatch

Batch insert edges into a Nebula database.

Functions

get_search_index_json

Get a JSON representation of a row for indexing in Meilisearch.

row_pk_hash

Compute a stable hash of a row’s primary key, and concatenate it with table name.