hoover3_database::models::collection
Module _nebula_edges

This module contains the definitions for all the nebula edges. Unit structs are used to identify edges, and are converted to GraphEdgeTypes for use at runtime.

Macros

declare_edge 🔒

Structs

FilesystemParentEdge

Unit struct to identify a Nebula edge `$id``.

Traits

GraphEdgeIdentifier

Trait for unit structs that can be used to identify a Nebula edge. These structs are to be used in code as a type safe identifier for an edge.

Functions

get_all_nebula_edge_types

Get all the Nebula edge types. Don’t forget to add new edges here…