hoover3_database::models::collection::_nebula_edges
Trait GraphEdgeIdentifier

pub trait GraphEdgeIdentifier: Sized {
    // Required method
    fn name(&self) -> DatabaseIdentifier;

    // Provided method
    fn to_owned(&self) -> GraphEdgeType { ... }
}

Trait for unit structs that can be used to identify a Nebula edge. These structs are to be used in code as a type safe identifier for an edge.

Required Methods

fn name(&self) -> DatabaseIdentifier

Get the name of the edge.

Provided Methods

fn to_owned(&self) -> GraphEdgeType

Convert the edge to a GraphEdgeType.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.