hoover3_database::models::collection
Module datasource

This module contains the table definitions for all the datasource related models.

Macros

delete_datasource_db_row 🔒

delete_datasource_db_row_query 🔒

find_datasource_db_row 🔒

find_datasource_db_row_query 🔒

find_first_datasource_db_row 🔒

partial_datasource_db_row 🔒

update_datasource_db_row_query 🔒

Structs

DatasourceDbRow

Database representation of a datasource in the system. This struct maps directly to a row in the datasources table on the collection keyspace.